pyproject.toml
[project]
name = "lanchain-tutorials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub>=0.33.2",
    "langchain-groq>=0.3.5",
    "langchain-huggingface>=0.3.0",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.27",
    "numpy>=2.3.1",
    "ollama>=0.5.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.0",
    "streamlit>=1.46.1",
    "torch>=2.7.1",
    "transformers>=4.53.1",
]
